Flatten GUI library dependencies; there should be no circular dependencies now
