+++ src/gui_error.h
