+++ src/gui_macros.h
