+++ src/gui_propmat.h
